more color tests
