+++ src/index.test.ts
@@ -4,64 +4,9 @@
  */
 
 import { assert } from 'chai';
-import { SixelDecoder, fromRGBA8888, toRGBA8888, sixelEncode } from './index';
+import { SixelDecoder, toRGBA8888, sixelEncode } from './index';
 import { RGBA8888 } from './Types';
 
-const BIG_ENDIAN = new Uint8Array(new Uint32Array([0xFF000000]).buffer)[0] === 0xFF;
-
-describe('RGBA8888 native colors', () => {
-  describe('toRGBA888', () => {
-    it('conversions', () => {
-      assert.equal(toRGBA8888(0, 0, 0, 0), 0);
-      assert.equal(toRGBA8888(0, 0, 0, 255), BIG_ENDIAN ? 0x000000FF : 0xFF000000);
-      assert.equal(toRGBA8888(0, 0, 255, 0), BIG_ENDIAN ? 0x0000FF00 : 0x00FF0000);
-      assert.equal(toRGBA8888(0, 255, 0, 0), BIG_ENDIAN ? 0x00FF0000 : 0x0000FF00);
-      assert.equal(toRGBA8888(255, 0, 0, 0), BIG_ENDIAN ? 0xFF000000 : 0x000000FF);
-    });
-    it('alpha defaults to 255', () => {
-      assert.equal(toRGBA8888(0, 0, 0), toRGBA8888(0, 0, 0, 255));
-      assert.equal(toRGBA8888(0, 0, 255), toRGBA8888(0, 0, 255, 255));
-      assert.equal(toRGBA8888(0, 255, 0), toRGBA8888(0, 255, 0, 255));
-      assert.equal(toRGBA8888(255, 0, 0), toRGBA8888(255, 0, 0, 255));
-      assert.equal(toRGBA8888(0, 255, 255), toRGBA8888(0, 255, 255, 255));
-      assert.equal(toRGBA8888(255, 0, 255), toRGBA8888(255, 0, 255, 255));
-      assert.equal(toRGBA8888(255, 255, 0), toRGBA8888(255, 255, 0, 255));
-      assert.equal(toRGBA8888(255, 255, 255), toRGBA8888(255, 255, 255, 255));
-    });
-    it('should only return unsigned', () => {
-      // test only for r and a here (g/b dont add to significant bit)
-      for (let r = 0; r <= 0xFF; ++r) {
-        for (let a = 0; a <= 0xFF; ++a) {
-          const color = toRGBA8888(r, 0, 0, a);
-          assert.equal(color >= 0, true);
-        }
-      }
-    });
-    it('handled signed channel values', () => {
-      assert.equal(toRGBA8888(-8, -50, -100, -127), toRGBA8888(-8 >>> 0, -50 >>> 0, -100 >>> 0, -127 >>> 0));
-    });
-    it('strip channel values to 8 bit (not clamping)', () => {
-      assert.equal(toRGBA8888(0x1234, 0x5678, 0xabcd, 0xef11), BIG_ENDIAN ? 0x3478cd11 : 0x11cd7834);
-    });
-  });
-  describe('fromRGBA8888', () => {
-    it('conversions', () => {
-      assert.deepEqual(fromRGBA8888(0), [0, 0, 0, 0]);
-      assert.deepEqual(fromRGBA8888(0x000000FF), BIG_ENDIAN ? [0, 0, 0, 255] : [255, 0, 0, 0]);
-      assert.deepEqual(fromRGBA8888(0x0000FF00), BIG_ENDIAN ? [0, 0, 255, 0] : [0, 255, 0, 0]);
-      assert.deepEqual(fromRGBA8888(0x00FF0000), BIG_ENDIAN ? [0, 255, 0, 0] : [0, 0, 255, 0]);
-      assert.deepEqual(fromRGBA8888(0xFF000000), BIG_ENDIAN ? [255, 0, 0, 0] : [0, 0, 0, 255]);
-    });
-    it('should only create unsigned channel values', () => {
-      assert.deepEqual(fromRGBA8888(-1), [255, 255, 255, 255]);
-      // 2 complement: -0xedcba988 ==> 0x12345678 (newDigit = 15 - digit; result + 1)
-      assert.deepEqual(fromRGBA8888(-0xedcba988), BIG_ENDIAN ? [0x12, 0x34, 0x56, 0x78] : [0x78, 0x56, 0x34, 0x12]);
-    });
-    it('strip values to 32bit', () => {
-      assert.deepEqual(fromRGBA8888(0x1234567890), BIG_ENDIAN ? [0x12, 0x34, 0x56, 0x78] : [0x90, 0x78, 0x56, 0x34])
-    });
-  });
-})
 
 describe('SixelDecoder', () => {
   let dec: SixelDecoder;
